Add options for the output (#48)
* Add options for the output

Add options controlling what shows in the output, by default this
follows the current output format, but adds an option to show the policy
name that provided the violation. Options are also added to remove the
display of the severity and the short explanation (PBP page).

* Add specs

Add some specs testing the output options and super basic testing of the
forced severity option.