ci(circle): apm clean after install
Newer version of apm use `npm prune` in the background, which with the 
newer backing version of npm actually installs packages in a completely 
blank node_modules environment. This can cause issues with some packages 
that expect certain parts to be installed already when they are 
uninstalled.

Switching the order shouldn't matter here and will prevent issues.